requirements.txt
google-cloud-storage==0.20.0